+++ Sources/TagKit/_Deprecated/TagCapsule.swift
@@ -3,15 +3,13 @@
 //  TagKit
 //
 //  Created by Daniel Saidi on 2022-08-19.
-//  Copyright © 2022-2024 Daniel Saidi. All rights reserved.
+//  Copyright © 2022-2025 Daniel Saidi. All rights reserved.
 //
 
+#if !os(watchOS)
 import SwiftUI
 
-/// This view can be used to render tags in a capsule shape.
-///
-/// This view will not slugify the provided tag string, only
-/// use the content you provide.
+@available(*, deprecated, message: "Just use a regular Text element and style it as you wish.")
 public struct TagCapsule: View {
 
     /// Create a tag capsule.
@@ -67,32 +65,7 @@ private extension View {
     }
 }
 
-#Preview {
-    
-    ZStack {
-        LinearGradient(colors: [.blue, .red], startPoint: .top, endPoint: .bottom)
-            .ignoresSafeArea()
-        
-        ScrollView {
-            VStack {
-                HStack {
-                    TagCapsule("standard")
-                        .tagCapsuleStyle(.standard)
-                    TagCapsule("standard-selected")
-                        .tagCapsuleStyle(.standardSelected)
-                }
-                HStack {
-                    TagCapsule("spider-man")
-                        .tagCapsuleStyle(.spiderman)
-                    TagCapsule("spider-man-selected")
-                        .tagCapsuleStyle(.spidermanSelected)
-                }
-            }
-            .padding(.top, 250)
-        }
-    }
-}
-
+@available(*, deprecated, message: "Just use a regular Text element and style it as you wish.")
 private extension TagCapsuleStyle {
     
     static var spiderman: Self {
@@ -120,3 +93,4 @@ private extension TagCapsuleStyle {
         return style
     }
 }
+#endif