+++ Sources/TagKit/_Deprecated/TagCapsuleStyle.swift
@@ -3,21 +3,13 @@
 //  TagKit
 //
 //  Created by Daniel Saidi on 2022-09-07.
-//  Copyright © 2022-2024 Daniel Saidi. All rights reserved.
+//  Copyright © 2022-2025 Daniel Saidi. All rights reserved.
 //
 
+#if !os(watchOS)
 import SwiftUI
 
-/// This style can be used to style ``TagCapsule`` views. It
-/// lets you define visual styles, like fore- and background
-/// colors, background materials, border, and padding.
-///
-/// To style other parts of the capsule, such as the font or
-/// the font weight, use native SwiftUI view modifiers.
-///
-/// The ``standard`` and ``standardSelected`` capsule styles
-/// use a primary foreground color, a clear background color
-/// and a material background.
+@available(*, deprecated, message: "Just use a regular Text element and style it as you wish.")
 public struct TagCapsuleStyle {
     
     /// Create a new tag capsule style.
@@ -94,6 +86,7 @@ public struct TagCapsuleStyle {
     public var padding: EdgeInsets
 }
 
+@available(*, deprecated, message: "Just use a regular Text element and style it as you wish.")
 public extension TagCapsuleStyle {
     
     struct Border : Sendable {
@@ -117,9 +110,6 @@ public extension TagCapsuleStyle {
         /// The border width.
         public var width: Double
     }
-}
-
-public extension TagCapsuleStyle {
     
     struct Shadow : Sendable{
         
@@ -154,9 +144,6 @@ public extension TagCapsuleStyle {
         /// The y offset.
         public var offsetY: Double
     }
-}
-
-public extension TagCapsuleStyle {
 
     /// The standard style.
     static var standard: Self {
@@ -173,6 +160,7 @@ public extension TagCapsuleStyle {
     }
 }
 
+@available(*, deprecated, message: "Just use a regular Text element and style it as you wish.")
 public extension TagCapsuleStyle.Shadow {
 
     /// The standard shadow style.
@@ -189,6 +177,7 @@ public extension TagCapsuleStyle.Shadow {
     }
 }
 
+@available(*, deprecated, message: "Just use a regular Text element and style it as you wish.")
 public extension TagCapsuleStyle.Border {
 
     /// The standard border style.
@@ -202,16 +191,7 @@ public extension TagCapsuleStyle.Border {
     }
 }
 
-public extension TagCapsuleStyle {
-    
-    @available(*, deprecated, renamed: "standard")
-    static var standardMaterial: TagCapsuleStyle { .standard }
-    
-    @available(*, deprecated, renamed: "standardSelected")
-    static var standardSelectedMaterial: TagCapsuleStyle { .standardSelected }
-}
-
-
+@available(*, deprecated, message: "Just use a regular Text element and style it as you wish.")
 public extension View {
 
     /// Apply a ``TagCapsule`` style to the view hierarchy.
@@ -222,6 +202,7 @@ public extension View {
     }
 }
 
+@available(*, deprecated, message: "Just use a regular Text element and style it as you wish.")
 private extension TagCapsuleStyle {
 
     struct Key: EnvironmentKey {
@@ -230,6 +211,7 @@ private extension TagCapsuleStyle {
     }
 }
 
+@available(*, deprecated, message: "Just use a regular Text element and style it as you wish.")
 public extension EnvironmentValues {
 
     /// This value can bind to a line spacing picker config.
@@ -238,3 +220,4 @@ public extension EnvironmentValues {
         set { self [TagCapsuleStyle.Key.self] = newValue }
     }
 }
+#endif